requirements.txt
graphics.py==5.0.1.post1
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
